Add files via upload
Connected the flights API. Now we can display basic folium.Markers for the airports and airplanes folium.Markers for the airplanes. The API is free and has no limit to the requests. Better find a new one because not all flights are displayed properly. Also, we need to change the airport CSV file because it is not updated.